+++ ImgwApi.Test/Hydro/HydroClientShould.cs
@@ -1,4 +1,4 @@
-﻿namespace ImgwApi.Test;
+﻿namespace ImgwApi.Test.Hydro;
 
 public class HydroClientShould
 {
