+++ ImgwApi/Hydro/HydroClient.cs
