+++ ImgwApi/Hydro/IHydroClient.cs
