+++ ImgwApi/Synop/ISynopClient.cs
