+++ ImgwApi/Synop/SynopClient.cs
