+++ ImgwApi/Synop/SynopData.cs
